+++ src/main/java/jinbok/culture/board/dto/BoardRequest.java
@@ -1,5 +1,7 @@
 package jinbok.culture.board.dto;
 
+import jakarta.validation.Valid;
+
 public record BoardRequest(
         String title,
         String content
