+++ src/main/java/jinbok/culture/user/dto/UserResponse.java
@@ -1,7 +1,5 @@
 package jinbok.culture.user.dto;
 
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotBlank;
 import jinbok.culture.user.domain.User;
 import lombok.Builder;
