+++ src/main/java/jinbok/culture/user/service/AuthService.java
@@ -39,4 +39,5 @@ public class AuthService {
                 .orElseThrow();
     }
 
+
 }
