+++ src/main/java/jinbok/culture/user/service/UserService.java
@@ -8,11 +8,22 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.GetMapping;
 
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class UserService {
 
+    public final UserRepository userRepository;
 
+    public UserResponse findByLoginId(Long id) {
+        return UserResponse.toUserResponse(userRepository.findById(id).orElseThrow());
+    }
+
+    public UserResponse updateUserInfo(Long id, UserRequest userRequest) {
+        User user = userRepository.findById(id).orElseThrow();
+
+
+    }
 }
